# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateJobTitleKategoriRequest:
    properties:
      code:
        maxLength: 255
        minLength: 1
        type: string
      fhir_code:
        maxLength: 500
        type: string
      label:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - code
    - label
    type: object
...
  dto.JobTitleKategoriResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/httputil.UserData'
      fhir_code:
        type: string
      id:
        type: integer
      label:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/httputil.UserData'
    type: object
  dto.JobTitleKategoriSimpelResponse:
    properties:
      code:
        type: string
      fhir_code:
        type: string
      id:
        type: integer
      label:
        type: string
    type: object
...
  dto.UpdateJobTitleKategoriRequest:
    properties:
      code:
        maxLength: 255
        minLength: 1
        type: string
      fhir_code:
        maxLength: 500
        type: string
      label:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - code
    - label
    type: object
...
  httputil.UserData:
    properties:
      id:
        type: integer
      name:
        type: string
      username:
        type: string
    type: object
...
  response.MyGoResponse:
    description: Standard JSON response
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
host: localhost:1323
info:
  contact:
    email: support@neosim.id
    name: Neosim Support
  description: REST API untuk Neosim — sistem manajemen informasi
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Neosim API V1
  version: "1.0"
paths:
...
  /kepegawaian/jabatan/job_title_kategoris:
    get:
      consumes:
      - application/json
      description: Get paginated list of JobTitleKategori
      parameters:
      - description: Filter by code (partial match)
        in: query
        name: code
        type: string
      - description: Filter by label (partial match)
        in: query
        name: label
        type: string
      - description: Sort column (code, label,fhir_code, created_at, updated_at)
        enum:
        - code
        - label
        - fhir_code
        - fhir_code
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.JobTitleKategoriResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get list of JobTitleKategori
      tags:
      - kepegawaian/jabatan/jabatan/job_title_kategori
    post:
      consumes:
      - application/json
      description: Create New JobTitleKategori
      parameters:
      - description: Create Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateJobTitleKategoriRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.JobTitleKategoriResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create JobTitleKategori
      tags:
      - kepegawaian/jabatan/jabatan/job_title_kategori
  /kepegawaian/jabatan/job_title_kategoris/{id}:
    delete:
      consumes:
      - application/json
      description: Delete JobTitleKategori by :id
      parameters:
      - description: JobTitleKategori ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MyGoResponse'
      security:
      - BearerAuth: []
      summary: Delete JobTitleKategori
      tags:
      - kepegawaian/jabatan/jabatan/job_title_kategori
    get:
      consumes:
      - application/json
      description: Get JobTitleKategori by :id
      parameters:
      - description: JobTitleKategori ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.JobTitleKategoriResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get JobTitleKategori
      tags:
      - kepegawaian/jabatan/jabatan/job_title_kategori
    put:
      consumes:
      - application/json
      description: Update JobTitleKategori by :id
      parameters:
      - description: JobTitleKategori ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateJobTitleKategoriRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.JobTitleKategoriResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update JobTitleKategori
      tags:
      - kepegawaian/jabatan/jabatan/job_title_kategori
  /kepegawaian/jabatan/job_title_kategoris/select:
    get:
      consumes:
      - application/json
      description: Get list of JobTitleKategori for select with optional search
      parameters:
      - description: Search by label or code
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.JobTitleKategoriSimpelResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get list of JobTitleKategori for select
      tags:
      - kepegawaian/jabatan/jabatan/job_title_kategori
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"