# internal/modules/kepegawaian/jabatan/handlers/job_title_kategori_handler.go
package handlers

import (
	"io"
	"net/http"

	"neosim_go/internal/modules/kepegawaian/jabatan/dto"
	"neosim_go/internal/shared/binding"
	he "neosim_go/internal/shared/httputil"
	"neosim_go/internal/shared/response"
	"neosim_go/internal/shared/validator"

	"github.com/labstack/echo/v5"
)

// Method di bawah ini ditempelkan ke struct KepegawaianJabatanHandler yang
// sama dengan handler entitas utama (lihat handlers/handler.go). Nama method
// diberi suffix JobTitleKategori agar tidak bentrok dengan method entitas utama
// pada struct handler yang sama.

// ─── ListSelectListJobTitleKategori ────────────────────────────────────────────
//
//	@Summary		Get list of JobTitleKategori for select
//	@Description	Get list of JobTitleKategori for select with optional search
//	@Tags			kepegawaian/jabatan/jabatan/job_title_kategori
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			search	query	string	false	"Search by label or code"
//	@Success		200		{object}	response.MyGoResponse{data=[]dto.JobTitleKategoriSimpelResponse}
//	@Router			/kepegawaian/jabatan/job_title_kategoris/select [get]
func (h *KepegawaianJabatanHandler) ListSelectListJobTitleKategori(c *echo.Context) error {
	search := c.QueryParam("search")

	actor := he.BuildAuthContext(c)
	items, err := h.service.ListSelectJobTitleKategori(c.Request().Context(), search, actor)
	if err != nil {
		return response.Response(c, http.StatusInternalServerError, false, "Gagal mengambil data : "+err.Error(), nil, nil)
	}
	return response.Response(c, http.StatusOK, true, "Berhasil mengambil data", items, nil)
}

// ─── ListJobTitleKategori ──────────────────────────────────────────────────────
//
//	@Summary		Get list of JobTitleKategori
//	@Description	Get paginated list of JobTitleKategori
//	@Tags			kepegawaian/jabatan/jabatan/job_title_kategori
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			code		query		string	false	"Filter by code (partial match)"
//	@Param			label		query		string	false	"Filter by label (partial match)"
//	@Param			sort_by		query		string	false	"Sort column (code, label,fhir_code, created_at, updated_at)"	Enums(code, label,fhir_code,fhir_code, created_at, updated_at)
//	@Param			sort_order	query		string	false	"Sort direction"	Enums(asc, desc)
//	@Param			page		query		int		false	"Page number"
//	@Param			page_size	query		int		false	"Page size"
//	@Success		200			{object}	response.MyGoResponse{data=[]dto.JobTitleKategoriResponse}
//	@Router			/kepegawaian/jabatan/job_title_kategoris [get]
func (h *KepegawaianJabatanHandler) ListJobTitleKategori(c *echo.Context) error {
	filter := dto.FilterJobTitleKategoriRequest{
		Code:  c.QueryParam("code"),
		Label: c.QueryParam("label"),
		// Sorting ---------------------------
		SortBy:    c.QueryParam("sort_by"),
		SortOrder: c.QueryParam("sort_order"),
	}
	page, pageSize := he.ParsePagination(c, h.cfg)

	actor := he.BuildAuthContext(c)
	items, total, err := h.service.ListJobTitleKategori(c.Request().Context(), page, pageSize, &filter, actor)
	if err != nil {
		return response.Response(c, http.StatusInternalServerError, false, err.Error(), nil, nil)
	}
	return response.Paginated(c, http.StatusOK, true, "Berhasil mengambil data", items, total, page, pageSize)
}

// ─── GetJobTitleKategoriByID ───────────────────────────────────────────────────
//
//	@Summary		Get JobTitleKategori
//	@Description	Get JobTitleKategori by :id
//	@Tags			kepegawaian/jabatan/jabatan/job_title_kategori
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			id	path		int	true	"JobTitleKategori ID"
//	@Success		200	{object}	response.MyGoResponse{data=dto.JobTitleKategoriResponse}
//	@Router			/kepegawaian/jabatan/job_title_kategoris/{id} [get]
func (h *KepegawaianJabatanHandler) GetJobTitleKategoriByID(c *echo.Context) error {
	id, err := he.ParseID(c)
	if err != nil {
		return response.Response(c, http.StatusBadRequest, false, "ID tidak valid", nil, nil)
	}
	actor := he.BuildAuthContext(c)
	item, err := h.service.GetJobTitleKategoriByID(c.Request().Context(), id, actor)
	if err != nil {
		return response.Response(c, http.StatusNotFound, false, err.Error(), nil, nil)
	}
	return response.Response(c, http.StatusOK, true, "Berhasil mengambil data", item, nil)
}

// ─── CreateJobTitleKategori ────────────────────────────────────────────────────
//
//	@Summary		Create JobTitleKategori
//	@Description	Create New JobTitleKategori
//	@Tags			kepegawaian/jabatan/jabatan/job_title_kategori
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			body	body		dto.CreateJobTitleKategoriRequest	true	"Create Request"
//	@Success		201		{object}	response.MyGoResponse{data=dto.JobTitleKategoriResponse}
//	@Router			/kepegawaian/jabatan/job_title_kategoris [post]
func (h *KepegawaianJabatanHandler) CreateJobTitleKategori(c *echo.Context) error {
	var req dto.CreateJobTitleKategoriRequest
	body, err := io.ReadAll(c.Request().Body)
	if err != nil {
		return response.Response(c, http.StatusBadRequest, false, "Gagal membaca request body", nil, err.Error())
	}

	if errs := binding.BindErrors(body, &req); len(errs) > 0 {
		return response.Response(c, http.StatusUnprocessableEntity, false, "Validasi gagal (binding)", nil, errs)
	}
	if errs := validator.Validate(req); errs != nil {
		return response.Response(c, http.StatusUnprocessableEntity, false, "Validasi gagal (validator)", nil, errs)
	}
	actor := he.BuildAuthContext(c)
	item, err := h.service.CreateJobTitleKategori(c.Request().Context(), &req, actor)
	if err != nil {
		return response.Response(c, http.StatusBadRequest, false, err.Error(), nil, nil)
	}
	return response.Response(c, http.StatusCreated, true, "Data berhasil dibuat", item, nil)
}

// ─── UpdateJobTitleKategori ────────────────────────────────────────────────────
//
//	@Summary		Update JobTitleKategori
//	@Description	Update JobTitleKategori by :id
//	@Tags			kepegawaian/jabatan/jabatan/job_title_kategori
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			id		path		int							true	"JobTitleKategori ID"
//	@Param			body	body		dto.UpdateJobTitleKategoriRequest	true	"Update Request"
//	@Success		200		{object}	response.MyGoResponse{data=dto.JobTitleKategoriResponse}
//	@Router			/kepegawaian/jabatan/job_title_kategoris/{id} [put]
func (h *KepegawaianJabatanHandler) UpdateJobTitleKategori(c *echo.Context) error {
	id, err := he.ParseID(c)
	if err != nil {
		return response.Response(c, http.StatusBadRequest, false, "ID tidak valid", nil, nil)
	}

	var req dto.UpdateJobTitleKategoriRequest
	body, err := io.ReadAll(c.Request().Body)
	if err != nil {
		return response.Response(c, http.StatusBadRequest, false, "Gagal membaca request body", nil, err.Error())
	}

	if errs := binding.BindErrors(body, &req); len(errs) > 0 {
		return response.Response(c, http.StatusUnprocessableEntity, false, "Validasi gagal (binding)", nil, errs)
	}
	if errs := validator.Validate(req); errs != nil {
		return response.Response(c, http.StatusUnprocessableEntity, false, "Validasi gagal (validator)", nil, errs)
	}
	actor := he.BuildAuthContext(c)
	item, err := h.service.UpdateJobTitleKategori(c.Request().Context(), id, &req, actor)
	if err != nil {
		status := http.StatusBadRequest
		if err.Error() == "JobTitleKategori tidak ditemukan" {
			status = http.StatusNotFound
		}
		return response.Response(c, status, false, err.Error(), nil, nil)
	}
	return response.Response(c, http.StatusOK, true, "Data berhasil diupdate", item, nil)
}

// ─── DeleteJobTitleKategori ────────────────────────────────────────────────────
//
//	@Summary		Delete JobTitleKategori
//	@Description	Delete JobTitleKategori by :id
//	@Tags			kepegawaian/jabatan/jabatan/job_title_kategori
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			id	path		int	true	"JobTitleKategori ID"
//	@Success		200	{object}	response.MyGoResponse{}
//	@Router			/kepegawaian/jabatan/job_title_kategoris/{id} [delete]
func (h *KepegawaianJabatanHandler) DeleteJobTitleKategori(c *echo.Context) error {
	id, err := he.ParseID(c)
	if err != nil {
		return response.Response(c, http.StatusBadRequest, false, "ID tidak valid", nil, nil)
	}
	actor := he.BuildAuthContext(c)
	if err := h.service.DeleteJobTitleKategori(c.Request().Context(), id, actor); err != nil {
		status := http.StatusInternalServerError
		if err.Error() == "JobTitleKategori tidak ditemukan" {
			status = http.StatusNotFound
		}
		return response.Response(c, status, false, err.Error(), nil, nil)
	}
	return response.Response(c, http.StatusOK, true, "Data berhasil dihapus", nil, nil)
}